+++ src/main/kotlin/ee/carlrobert/codegpt/settings/models/ModelSettings.kt
@@ -5,6 +5,7 @@ import com.intellij.openapi.components.*
 import ee.carlrobert.codegpt.settings.service.FeatureType
 import ee.carlrobert.codegpt.settings.service.ModelChangeNotifier
 import ee.carlrobert.codegpt.settings.service.ServiceType
+import ee.carlrobert.codegpt.settings.service.ServiceType.PROXYAI
 import ee.carlrobert.codegpt.settings.service.custom.CustomServicesSettings
 
 @Service
@@ -61,6 +62,7 @@ class ModelSettings : SimplePersistentStateComponent<ModelSettingsState>(ModelSe
         migrateCustomOpenAIModelCodesToIds()
         migrateMissingProviderInformation()
         migrateEditCodeModel()
+        migrateProxyAIModels()
         notifyIfChanged(oldState, this.state)
     }
 
@@ -141,6 +143,15 @@ class ModelSettings : SimplePersistentStateComponent<ModelSettingsState>(ModelSe
         state.modelSelections["EDIT_CODE"]?.let {
             state.setModelSelection(FeatureType.INLINE_EDIT, it.model, it.provider!!)
             state.modelSelections.remove("EDIT_CODE")
+        }
+    }
+
+    private fun migrateProxyAIModels() {
+        listOf(FeatureType.AUTO_APPLY, FeatureType.CODE_COMPLETION, FeatureType.NEXT_EDIT).forEach {
+            val modelSelection = state.getModelSelection(it)
+            if (modelSelection?.provider == PROXYAI && modelSelection.model != ModelRegistry.MERCURY_CODER) {
+                setModelWithProvider(it, ModelRegistry.MERCURY_CODER, PROXYAI)
+            }
         }
     }