+++ src/main/kotlin/ee/carlrobert/codegpt/toolwindow/chat/editor/SearchReplaceHandler.kt
@@ -4,6 +4,7 @@ import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.editor.EditorKind
 import com.intellij.openapi.editor.ex.EditorEx
 import com.intellij.openapi.vfs.readText
+import ee.carlrobert.codegpt.util.EditorUtil
 import ee.carlrobert.codegpt.CodeGPTKeys
 import ee.carlrobert.codegpt.codecompletions.CompletionProgressNotifier
 import ee.carlrobert.codegpt.toolwindow.chat.editor.ResponseEditorPanel.Companion.RESPONSE_EDITOR_DIFF_VIEWER_VALUE_PAIR_KEY
@@ -60,7 +61,7 @@ class SearchReplaceHandler(
             return
         }
 
-        val currentText = virtualFile.readText()
+        val currentText = EditorUtil.getFileContent(virtualFile)
         val containsText = currentText.contains(searchContent.trim())
 
         if (searchContent.isNotEmpty() && editor.editorKind == EditorKind.DIFF && !containsText && !searchFailed) {