+++ src/main/kotlin/ee/carlrobert/codegpt/toolwindow/chat/editor/actions/AutoApplyAction.kt
@@ -12,6 +12,9 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.readText
 import com.intellij.ui.components.AnActionLink
 import com.intellij.util.ui.JBUI
+import ee.carlrobert.codegpt.settings.service.FeatureType
+import ee.carlrobert.codegpt.settings.service.ModelSelectionService
+import ee.carlrobert.codegpt.settings.service.ServiceType.INCEPTION
 import ee.carlrobert.codegpt.util.EditorUtil
 import javax.swing.JComponent
 
@@ -38,7 +41,7 @@ class AutoApplyAction(
             anActionLink.isEnabled = true
             anActionLink.toolTipText = "Apply changes to ${virtualFile.name}"
 
-            if (virtualFile.readText().trim() == toolwindowEditor.document.text.trim()) {
+            if (EditorUtil.getFileContent(virtualFile).trim() == toolwindowEditor.document.text.trim()) {
                 anActionLink.isEnabled = false
                 anActionLink.isVisible = true
                 anActionLink.toolTipText = "No changes to apply"